template/Django/requirements.txt
asgiref
Django
django-dotenv
django-environ
djangorestframework
pytz
sqlparse
tzdata
